Setting default schema values in FormEditor
